requirements.txt
easygui==0.98.2
opencv-python==4.5.1.48
Pillow==7.2.0